pyproject.toml
[project]
name = "llm-economy-summary"
version = "0.1.0"
description = ""
authors = [
    {name = "runru1030",email = "runru1030@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "fastapi (>=0.127.0,<0.128.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "psycopg2-binary (>=2.9.0,<3.0.0)",
    "asyncpg (>=0.29.0,<0.30.0)",
    "greenlet (>=3.3.0,<4.0.0)",
    "langchain (>=1.2.6,<2.0.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "langchain-openai (>=1.1.7,<2.0.0)",
    "langgraph (>=1.0.6,<2.0.0)",
    "chromadb (>=1.4.1,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "langchain-chroma (>=1.1.0,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "sse-starlette (>=3.2.0,<4.0.0)",
    "langgraph-checkpoint-postgres (>=3.0.3,<4.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "langfuse (>=3.12.0,<4.0.0)",
    "ruff (>=0.14.13,<0.15.0)"
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "A",
    "ARG",
    "B",
    "C",
    "DTZ",
    "E",
    "EM",
    "F",
    "FBT",
    "I",
    "ICN",
    "ISC",
    "N",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "Q",
    "RUF",
    "S",
    "T",
    "UP",
    "W",
    "YTT",
]
ignore = [
    # Allow non-abstract empty methods in abstract base classes
    "B027",
    # Allow boolean positional values in function calls, like `dict.get(... True)`
    "FBT003",
    # Ignore checks for possible passwords
    "S105",
    "S106",
    "S107",
    # Ignore complexity
    "C901",
    "PLR0911",
    "PLR0912",
    "PLR0913",
    "PLR0915",
    # Ignore Except
    "E722",
    "EM101",
    "EM102",
    # boolean-type-hint-positional-argument
    "FBT001",
    "FBT002",
    # unused-function-argument
    "ARG001",
    "S104",
    "B904",
    "PLR2004",
    "A003",
    "ARG002",
    "A005",
]
unfixable = [
    # Don't touch unused imports
    "F401",
]
extend-ignore = ["ISC001"]

[tool.ruff.lint.isort]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.per-file-ignores]
# Tests can use magic values, assertions, and relative imports
"test/**/*" = [
    "PLR2004",
    "PLC2401",
    "S101",
    "TID252",
    "S311",
    "ARG002",
    "T201",
    "N802",
    "E402",
]
"**/*_test.py" = [
    "PLR2004",
    "PLC2401",
    "S101",
    "TID252",
    "S311",
    "ARG002",
    "T201",
    "N802",
    "E402",
]
"**/*/__init__.py" = ["F401"]
"migration/**" = ["F401"]